// webapp/src/app/shared/models/permission.ts
export enum Permission {
    SuperAdmin = 'super-admin',
    Admin = 'admin',
    UserReadOnly = 'user-readonly',
    EventAdmin = 'event-admin',
    EventReadOnly = 'event-readonly',
    CourseAdmin = 'course-admin',
    CourseReadOnly = 'course-readonly',
    LibraryAdmin = 'library-admin',
    LibraryReadOnly = 'library-readonly',
    TransactionAdmin = 'transaction-admin',
    TransactionReadOnly = 'transaction-readonly',
    SchoolAdmin = 'school-admin',
    SchoolReadOnly = 'school-readonly',
  }
  